Actually support simultaneous requests from multiple clients (partly implemented, query_request still needs these changes)
When multiple devices send requests, Node uses the same "thread"/"process" to handle the response. My old implementation
assumed that each request got a separate thread, when in reality, simultaneous requests (or requests really close to each other)
were using the same global variables to create a response (very very bad, often resulted in crashes or a request getting a response
meant for another request)

client_sync now uses associated arrays as global variables instead. When a request comes in, a unique key is assigned to the request
and this key is passed throughout the response creation. This key is used to access the global array, returning global information
pertaining to that request, and that request only.

Same changes need to be made in query_request